+++ Engine/Resources/MainResources.xaml.cs
@@ -269,6 +269,8 @@ namespace JocysCom.VS.AiCompanion.Engine.Resources
         public static string main_Max_Output_Tokens => FindResource(nameof(main_Max_Output_Tokens));
         public static string main_Max_Output_Tokens_Help => FindResource(nameof(main_Max_Output_Tokens_Help));
         public static string main_Model => FindResource(nameof(main_Model));
+        public static string main_Model_Instructions => FindResource(nameof(main_Model_Instructions));
+        public static string main_Model_Instructions_Help => FindResource(nameof(main_Model_Instructions_Help));
         public static string main_Features_are_kown => FindResource(nameof(main_Features_are_kown));
         public static string main_ReasoningEffort => FindResource(nameof(main_ReasoningEffort));
         public static string main_ReasoningEffort_Help => FindResource(nameof(main_ReasoningEffort_Help));