+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-2025-01-13 v1.16.22
+2025-01-13 v1.16.24
 
 - New: `GenerateQrCodeImage` function. Can be used to send data to a mobile using the camera.
 - New: `ReasoningEffort` option for the `o1` and `o1-mini` reasoning models.
+- New: The option to provide specific instructions for each AI model.
 
 2025-01-12 v1.16.20